refactor: move webhooks into cli
